temporarily change libomptfinalizer to a singleton rather than a list
the debugger showed that the list was corrupted when the finalizer was
called. it obscured the locations of the list contents, so a watchpoint
couldn't be used to find the problem. change it to a singleton
that can be monitored. this worked. add an assert that will trip
if anyone registers a second finalizer. it is an adequate, though
unsatisfying patch.